% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{action_bar}
\alias{action_bar}
\title{Generate a horizontal action bar (flex row of buttons/links)}
\usage{
action_bar(..., justify = "flex-start")
}
\arguments{
\item{...}{action_button(), modal_link(), or other HTML strings}

\item{justify}{CSS justify-content value (default "flex-start")}
}
\value{
HTML string
}
\description{
Generate a horizontal action bar (flex row of buttons/links)
}
