% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{badge}
\alias{badge}
\title{Generate a styled badge}
\usage{
badge(text, class = NULL, data_value = NULL, style = list())
}
\arguments{
\item{text}{Badge text (will be HTML-escaped)}

\item{class}{CSS class(es) for color coding (e.g., "eng-identifierad")}

\item{data_value}{Optional data-value attribute for CSS targeting}

\item{style}{Named list of additional inline CSS}
}
\value{
HTML string
}
\description{
Generate a styled badge
}
