% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{detail_grid}
\alias{detail_grid}
\title{Generate a grid layout for detail sections}
\usage{
detail_grid(..., cols = 2)
}
\arguments{
\item{...}{detail_row() outputs or other HTML strings}

\item{cols}{Number of columns (default 2)}
}
\value{
HTML string
}
\description{
Generate a grid layout for detail sections
}
