% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{modal_link}
\alias{modal_link}
\title{Generate a link that opens a modal}
\usage{
modal_link(
  label,
  modal_id,
  params = list(),
  size = "md",
  style = list(),
  class = NULL,
  icon = NULL
)
}
\arguments{
\item{label}{Link text (will be HTML-escaped)}

\item{modal_id}{Modal identifier (e.g., "contacts/profile")}

\item{params}{Named list of query parameters}

\item{size}{Modal size: "sm", "md", "lg", "xl"}

\item{style}{Named list of inline CSS properties}

\item{class}{Additional CSS class(es)}

\item{icon}{Bootstrap Icon name}
}
\value{
HTML string
}
\description{
Generate a link that opens a modal
}
